Add human panics

See merge request mkonig/fastpasta!141

+++ fastpasta/src/stats/stats_collector/rdh_stats.rs
@@ -181,7 +181,7 @@ impl RdhStats {
         self.rdhs_seen += 1;
     }
 
-    pub(super) fn add_rdhs_seen(&mut self, rdhs_seen: u16) {
+    pub(super) fn add_rdhs_seen(&mut self, rdhs_seen: u32) {
         self.rdhs_seen += rdhs_seen as u64;
     }
 
@@ -194,7 +194,7 @@ impl RdhStats {
         self.rdhs_filtered += 1;
     }
 
-    pub(super) fn add_rdhs_filtered(&mut self, rdhs_filtered: u16) {
+    pub(super) fn add_rdhs_filtered(&mut self, rdhs_filtered: u32) {
         self.rdhs_filtered += rdhs_filtered as u64;
     }
